Separate binary_op module
