+++ src/data/value/binary_op/i64.rs
@@ -1,0 +1,95 @@
+use {
+    super::TryBinaryOperator,
+    crate::{data::ValueError, prelude::Value, result::Result},
+    std::cmp::Ordering,
+    Value::*,
+};
+
+impl PartialOrd<Value> for i64 {
+    fn partial_cmp(&self, rhs: &Value) -> Option<Ordering> {
+        match rhs {
+            I8(rhs) => PartialOrd::partial_cmp(self, &(*rhs as i64)),
+            I64(rhs) => PartialOrd::partial_cmp(self, rhs),
+            F64(rhs) => PartialOrd::partial_cmp(&(*self as f64), rhs),
+            _ => None,
+        }
+    }
+}
+
+impl TryBinaryOperator for i64 {
+    type Rhs = Value;
+
+    fn try_add(&self, rhs: &Self::Rhs) -> Result<Value> {
+        let lhs = *self;
+
+        match *rhs {
+            I8(rhs) => Ok(I64(lhs + rhs as i64)),
+            I64(rhs) => Ok(I64(lhs + rhs)),
+            F64(rhs) => Ok(F64(lhs as f64 + rhs)),
+            Null => Ok(Null),
+            _ => Err(ValueError::AddOnNonNumeric(I64(lhs), rhs.clone()).into()),
+        }
+    }
+
+    fn try_subtract(&self, rhs: &Self::Rhs) -> Result<Value> {
+        let lhs = *self;
+
+        match *rhs {
+            I8(rhs) => Ok(I64(lhs - rhs as i64)),
+            I64(rhs) => Ok(I64(lhs - rhs)),
+            F64(rhs) => Ok(F64(lhs as f64 - rhs)),
+            Null => Ok(Null),
+            _ => Err(ValueError::SubtractOnNonNumeric(I64(lhs), rhs.clone()).into()),
+        }
+    }
+
+    fn try_multiply(&self, rhs: &Self::Rhs) -> Result<Value> {
+        let lhs = *self;
+
+        match *rhs {
+            I8(rhs) => Ok(I64(lhs * rhs as i64)),
+            I64(rhs) => Ok(I64(lhs * rhs)),
+            F64(rhs) => Ok(F64(lhs as f64 * rhs)),
+            Interval(rhs) => Ok(Interval(lhs * rhs)),
+            Null => Ok(Null),
+            _ => Err(ValueError::MultiplyOnNonNumeric(I64(lhs), rhs.clone()).into()),
+        }
+    }
+
+    fn try_divide(&self, rhs: &Self::Rhs) -> Result<Value> {
+        let lhs = *self;
+
+        match *rhs {
+            I8(rhs) => Ok(I64(lhs / rhs as i64)),
+            I64(rhs) => Ok(I64(lhs / rhs)),
+            F64(rhs) => Ok(F64(lhs as f64 / rhs)),
+            Null => Ok(Null),
+            _ => Err(ValueError::DivideOnNonNumeric(I64(lhs), rhs.clone()).into()),
+        }
+    }
+
+    fn try_modulo(&self, rhs: &Self::Rhs) -> Result<Value> {
+        let lhs = *self;
+
+        match *rhs {
+            I8(rhs) => Ok(I64(lhs % rhs as i64)),
+            I64(rhs) => Ok(I64(lhs % rhs)),
+            F64(rhs) => Ok(F64(lhs as f64 % rhs)),
+            Null => Ok(Null),
+            _ => Err(ValueError::ModuloOnNonNumeric(I64(lhs), rhs.clone()).into()),
+        }
+    }
+}
+
+impl PartialEq<Value> for f64 {
+    fn eq(&self, other: &Value) -> bool {
+        let lhs = *self;
+
+        match *other {
+            I8(rhs) => lhs == rhs as f64,
+            I64(rhs) => lhs == rhs as f64,
+            F64(rhs) => lhs == rhs,
+            _ => false,
+        }
+    }
+}
